Hide dev settings section for non-debug builds. Move XRD test tokens to Account settings screen
